AIRA/Renal Web App/requirements.txt
dicom2nifti==2.4.3
Flask==1.1.2
matplotlib==3.5.1
MySQL-python==1.2.5
nibabel==4.0.2
numpy==1.21.5
opencv_python==4.6.0.66
patool==1.12
Pillow==9.2.0
scikit_learn==1.1.3
segmentation_models==1.0.1
tensorflow==2.9.0
tqdm==4.64.0
Werkzeug==2.0.3